backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
google-generativeai
python-dotenv
pytest
